requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase>=2.0.0
PyJWT>=2.8.0
flasgger>=0.9.7